[fixed] TC fixes
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@1132 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/tangible/deed/harvesterdeed/HarvesterDeedImplementation.cpp
@@ -40,9 +40,9 @@ HarvesterDeedImplementation::HarvesterDeedImplementation(CreatureObject* creatur
 	info(templateName);
 	name = "";
 
-	setSurplusMaintenance(hino->getSurplusMaintenance());
+	setSurplusMaintenance((uint32)hino->getSurplusMaintenance());
 	setMaintenanceRate(hino->getMaintenanceRate());
-	setSurplusPower(hino->getSurplusPower());
+	setSurplusPower((uint32)hino->getSurplusPower());
 	setPowerRate(hino->getPowerRate());
 	setHopperSize(hino->getHopperSizeMax());
 	setExtractionRate(hino->getExtractionRate());
